✨feat: Account 전체 클래스 작성 및 수정

+++ backend/src/main/java/com/dream/backend/domain/bank/Bank.java
@@ -16,7 +16,7 @@ public class Bank {
     @Id
     @GeneratedValue()
     @Column(name=" bank_code")
-    private int bankCode;
+    private Long bankCode;
 
     @Column(nullable = false, length = 50)
     private String bankName;
